+++ draw.h
@@ -6,6 +6,6 @@
 size_t calcBitmapSize(size_t width, size_t height);
 bool createBitmap(size_t width, size_t height);
 bool saveBitmap(char* filename);
-void setPixel(int x, int y, uint8_t color, float fsub);
+void setPixel(size_t x, size_t y, uint8_t color, float fsub);
 
 #endif
